Added demonstration of depth limit.
The MiniMax algorithm prvides a depth limit for its search for the best move if tht limit is reached the algortihm must estimate the value of a given board position. Our estimator counts how many ways to win a player has whoich arent blocked by the other player and returns the difference between these estimates as the value of the board position.