allow soft match only when moving a pawn or king
There was a bug because we were using the soft match all the time.
This led to situations where moving a rook could cause castling.
Now we only allow soft matches when a pawn is moving or a king is
moving. The soft match brings in the extra effects of castling or
en-passant. Since these moves will never be ambiguous with any of
the exact match moves this is a fine thing to do.